refactor extra visualizations
